standard exceptions

+++ db/db_user.py
@@ -3,6 +3,8 @@ from db.hash import Hash
 from schemas import UserBase
 from db.models import DbUser
 
+from fastapi import HTTPException,status
+
 
 def create_user(db : Session, request : UserBase):
     new_user = DbUser(
@@ -19,10 +21,18 @@ def get_all_users(db:Session):
     return db.query(DbUser).all()# read all users from table
 
 def get_user(db:Session, id:int):
-    return db.query(DbUser).filter(DbUser.id==id).first()
+    user = db.query(DbUser).filter(DbUser.id==id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail =f'User with id {id} not found')
+    return user
+
 
 def update_user(db:Session,id:int,request:UserBase):
     user = db.query(DbUser).filter(DbUser.id==id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail =f'User with id {id} not found')
     user.update({DbUser.username:request.username,
                  DbUser.email:request.email,
                  DbUser.password:Hash.brcypt(request.password)})
@@ -31,6 +41,9 @@ def update_user(db:Session,id:int,request:UserBase):
 
 def delete_user(db:Session,id :int):
     user = db.query(DbUser).filter(DbUser.id==id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail =f'User with id {id} not found')
     db.delete(user)
     db.commit()
     return 'ok deleted'
